backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.5.0
Pillow==11.1.0
